Moving from cgb to wordpress-scripts
