--- .eslintignore
@@ -1,9 +1,0 @@
-**/*.min.js
-**/*.build.js
-**/node_modules/**
-**/vendor/**
-build
-coverage
-cypress
-node_modules
-vendor
