+++ block.json
@@ -1,0 +1,14 @@
+{
+	"name": "wp-gutenberg-epfl/blocks",
+	"title": "EPFL Gutenberg Blocks",
+	"category": "common",
+	"icon": "",
+	"description": "",
+	"textdomain": "wp-gutenberg-epfl",
+	"supports": {
+		"html": false
+	},
+	"editorScript": "file:./build/index.js",
+	"editorStyle": "file:./build/index.css",
+	"style": "file:./build/style-index.css"
+}
