+++ src/editor.scss
@@ -1,0 +1,9 @@
+/**
+ * The following styles get applied inside the editor only.
+ *
+ * Replace them with your own styles or remove the file completely.
+ */
+
+.wp-block-wp-gutenberg-epfl-wp-gutenberg-epfl {
+	border: 1px dotted #f00;
+}
