+++ src/style.scss
@@ -1,0 +1,79 @@
+/**
+ * The following styles get applied both on the front of your site
+ * and in the editor.
+ *
+ * Replace them with your own styles or remove the file completely.
+ */
+/**
+ * #.# Common SCSS
+ *
+ * Can include things like variables and mixins
+ * that are used across the project.
+*/
+
+// Colors.
+$black: rgb(41, 41, 41);
+$white: #f4f4f4;
+$gray: #dedede;
+$green: #bada55;
+$red: orangered;
+
+#preview-box {
+	background-color: lightgrey;
+	border-color: black;
+	border-style: solid;
+	border-width: 1px;
+	margin: 2px 4px;
+}
+
+#preview-box h2 {
+	text-align: center;
+}
+
+#preview-box .helper {
+	text-align: center;
+	margin-bottom: 25px;
+}
+
+.block-editor-block-list__block .block-editor-block-list__block-edit {
+	margin-left: -28px;
+	margin-right: -28px;
+	border: 1px #e2e4e7 solid;
+	padding: 16px;
+}
+
+.wp-block-scroll {
+	-webkit-appearance: none;  // for MacOS trackpad users
+	max-height: 398px;
+	overflow-y: scroll;
+	overflow-x: hidden;
+	padding-right: 12px;
+	margin-right: -16px;
+	margin-bottom: -4px;
+}
+
+.wp-block-help {
+	margin-left: 44px;
+}
+
+.field-with-no-margin-bottom {
+	margin-bottom: 0 !important;
+}
+
+.epfl-uploader-image-wrapper {
+	position: relative;
+}
+
+.epfl-uploader-img {
+	width: 100%;
+}
+.epfl-uploader-remove-image {
+	position: absolute;
+	top: 10px;
+	right: 10px;
+	color: black;
+	border-width: 1px;
+	background-color: white;
+	color: #191e23;
+	box-shadow: inset 0 0 0 1px #e2e4e7, inset 0 0 0 2px #fff, 0 1px 1px rgba(25,30,35,.2);
+}
